Begin work on a proper AST
Refs #7

When I first started this project, it was in Python, and was
an experiment in finding the javadoc for methods, vars, etc.
Since it was so focused, I thought I could skip a lot of hard
parsing to gain speed; because it was an experiment, I thought
I could use regex. Well, both suspicious were wrong; I started
adding AST-esque parsing for better accuracy, but ended scrapping
the project because the neither the performance nor the usefulness
were there.

For Njast, I abandoned regex but was still thinking mostly about
Javadoc, and was hopeful that I could skip handling a lot of
complicated stuff and focus on what I needed, so I tried to
parse Java *semantically*, but in the AST mindset (hence the name).
This worked quite well for a while, but when I decided I wanted
Tern/Jedi-style autocomplete, I started to discover that my
half-assed AST just wasn't working, and was leading towards
horrible code that was hard to understand, much less maintain.
Worse, it just couldn't provide the type if information I found
myself wanting.

And so begins my adventure in proper AST parsing. I learned a lot
writing the AST for Robocon, but that has neither control flow nor
comments, so this should be interesting.