Fix resolveDeclaringType and have evaluateType depend on it
All projections now take a callback so types can be filled out.
This is likely the cause of the decreased performance (~115ms
for the ~100 ast tests to ~155ms). Also, we currently don't
cache specific (ex: {method: 'name'}) projections from openClass,
which might be causing unnecessary IO. We could/should consider caching
these projections to reduce that.

We need a way to invalidate caches, however, and we should also
pro-actively fill caches for the vim plugin---there's plenty of time
at startup for us to at least parse the current AST, and possibly
pre-cache parent types. This would, of course, depend on the better
cache handling mentioned above