Prepare to support eager caching via "init" call
Pretty much required for large files (> 1000 lines, probably); otherwise
the perf is just tragically poor. We will also need to be able to send
partial buffers---possibly the current block, although special
consideration will be needed when the cursor is in a ClassBody....

One possibility is to just search for a block boundary. *Probably*
the relevent new context would be above the cursor, but we may
not be able to rely on that....