Experiment with async handling missing imports; minor fixes
We don't actually do anything yet, but the stuff in place seems
to work. That said, "init" actually runs very fast, so it may
be feasible to just do it synchronously. I guess we'll have to see
how slow searching for the right import will be.... We may want
to consider separating this to a separate bufWrite endpoint, so
init can pre-cache/index stuff, and bufWrite can do the import
searching.